Adds unit happy paths of GameServiceImpl.

+++ src/main/java/com/minesweeper/api/models/Game.java
@@ -20,9 +20,9 @@ public class Game implements Serializable {
   @JsonProperty("_id")
   String id;
   String userEmail;
-  int rows;
-  int columns;
-  int mines;
+  long rows;
+  long columns;
+  long mines;
   Map<String, Cell> cells;
   Date startDateMillis;
   Date resumeDateMillis;
